Extract new spec ExtImport mapper
