+++ waspc/data/packages/wasp-config/src/spec/extImport.ts
@@ -1,0 +1,34 @@
+import type * as AppSpec from "../appSpec.js";
+
+export type ExtImport = NamedExtImport | DefaultExtImport;
+
+export interface NamedExtImport {
+  import: string;
+  alias?: string;
+  from: AppSpec.ExtImport["path"];
+}
+
+export interface DefaultExtImport {
+  importDefault: string;
+  from: AppSpec.ExtImport["path"];
+}
+
+export function mapExtImport(extImport: ExtImport): AppSpec.ExtImport {
+  if ("import" in extImport) {
+    return {
+      kind: "named",
+      name: extImport.import,
+      path: extImport.from,
+    };
+  } else if ("importDefault" in extImport) {
+    return {
+      kind: "default",
+      name: extImport.importDefault,
+      path: extImport.from,
+    };
+  } else {
+    throw new Error(
+      "Invalid ExtImport: neither `import` nor `importDefault` is defined",
+    );
+  }
+}
